fix some error
